+++ vins_estimator/src/utility/utility.h
@@ -152,3 +152,5 @@ class Utility
             two_pi * std::floor((-angle_degrees + T(180)) / two_pi);
     };
 };
+
+
